requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
timm>=0.9.0
numpy>=1.24.0
tqdm>=4.65.0
streamlit>=1.34.0
watchdog>=3.0.0
python-dotenv>=1.0.0 